Updated UI for all models

+++ modelos_filas_project/simulator/views.py
@@ -13,21 +13,21 @@ def input(request):
 
 def mm1Results(request):
     name_model = request.GET.get('name_model')
-    data = m.Model(name_model, int(request.GET.get('l')), 
-        int(request.GET.get('miu')), int(request.GET.get('n')), request.GET.get('time'))
+    data = m.Model(name_model, float(request.GET.get('l')), 
+        float(request.GET.get('miu')), int(request.GET.get('n')), request.GET.get('time'))
     results = {'model': data.model, 'name_model': name_model}
     return render(request, 'simulator/mm1_results.html', results)
 
 def mmsResults(request):
     name_model = request.GET.get('name_model')
-    data = m.Model(name_model, int(request.GET.get('l')), int(request.GET.get('miu')), 
+    data = m.Model(name_model, float(request.GET.get('l')), float(request.GET.get('miu')), 
     int(request.GET.get('s')), int(request.GET.get('n')), request.GET.get('time'))
     results = {'model': data.model, 'name_model': name_model}
     return render(request, 'simulator/mms_results.html', results)
 
 def mmskResults(request):
     name_model = request.GET.get('name_model')
-    data = m.Model(name_model, int(request.GET.get('l')), int(request.GET.get('miu')), 
+    data = m.Model(name_model, float(request.GET.get('l')), float(request.GET.get('miu')), 
     int(request.GET.get('s')), int(request.GET.get('k')), int(request.GET.get('n')),
     request.GET.get('time'))
     results = {'model': data.model, 'name_model': name_model}
@@ -35,8 +35,8 @@ def mmskResults(request):
 
 def mg1Results(request):
     name_model = request.GET.get('name_model')
-    data = m.Model(name_model, int(request.GET.get('l')), 
-    int(request.GET.get('miu')), int(request.GET.get('variance')), 
+    data = m.Model(name_model, float(request.GET.get('l')), 
+    float(request.GET.get('miu')), float(request.GET.get('variance')), 
     int(request.GET.get('n')), request.GET.get('time'))
     results = {'model': data.model, 'name_model': name_model}
     return render(request, 'simulator/mg1_results.html', results)
